sources/requirements.txt
pika==1.3.0
pymongo==4.2.0
paho-mqtt==1.6.1
typer==0.6.1
colorama==0.4.5
shellingham==1.5.0
